nwb_linkml/pyproject.toml
[project]
name = "nwb-linkml"
description = "Translating NWB schema language to LinkML"
version = "0.2.0"
authors = [
    {name = "sneakers-the-rat", email = "sneakers-the-rat@protonmail.me"},
]
license = {text = "AGPL-3.0"}
readme = "README.md"
requires-python = "<3.13,>=3.10"
dependencies = [
    "pyyaml>=6.0",
    "linkml-runtime>=1.7.7",
    "nwb-schema-language>=0.1.3",
    "rich>=13.5.2",
    #"linkml>=1.7.10",
    "linkml @ git+https://github.com/sneakers-the-rat/linkml@nwb-linkml",
    "pydantic>=2.3.0",
    "h5py>=3.9.0",
    "pydantic-settings>=2.0.3",
    "tqdm>=4.66.1",
    'typing-extensions>=4.12.2;python_version<"3.11"',
    "numpydantic>=1.3.0",
    "black>=24.4.2",
    "pandas>=2.2.2",
]

[project.urls]
homepage = "https://nwb-linkml.readthedocs.io"
repository = "https://github.com/p2p-ld/nwb-linkml/"
documentation = "https://nwb-linkml.readthedocs.io"

[project.optional-dependencies]
plot = [
    "dash<3.0.0,>=2.12.1",
    "dash-cytoscape<1.0.0,>=0.3.0",
]
tests = [
    "nwb-linkml[plot]",
    "pytest<8.0.0,>=7.4.0",
    "pytest-depends<2.0.0,>=1.0.1",
    "coverage<7.0.0,>=6.1.1",
    "pytest-md<1.0.0,>=0.2.0",
    "pytest-cov<5.0.0,>=4.1.0",
    "coveralls<4.0.0,>=3.3.1",
    "pytest-profiling<2.0.0,>=1.7.0",
    "sybil<6.0.0,>=5.0.3",
    "requests-cache>=1.2.1",
]
dev = [
    "nwb-linkml[tests]",
    "ruff>=0.5.0",
    "black>=24.4.2",
]

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src/nwb_linkml", "src/nwb_linkml/schema/**/*"]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
addopts = [
    "--cov=nwb_linkml",
    "--cov-append",
    "--cov-config=.coveragerc",
    "-p no:doctest",
    "--ignore=tests/__tmp__",
    "--ignore=src/nwb_linkml/models",
    "--ignore=src/nwb_linkml/schema"
]
markers = [
    "dev: tests that are just for development rather than testing correctness",
    "provider: tests for providers!"
]
testpaths = [
    "src/nwb_linkml",
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:parse_obj:pydantic.PydanticDeprecatedSince20"
]
